Wrote a 3d6 probability function
wrote a function that calculates the probability of getting any result for rolling any number of d6 dice. I could add another parameter that makes it any number of dx dice, but for now I decided to bake the 6 into the function, should be easy to change.

additionally, the result that I'm getting is correct...ish. There are rampant floating point errors, and 1. I don't want the errors adding up if I keep using my functions in functions, and 2. it looks ugly, so I'll fix that next time.