Calculated average DPS for the first time!
Wrote a function that assumes a regular gaming dice and calculates the average roll of that dice. Wrote another function that calculates Damage Per Second (DPS). We identified that a textbook GURPS scenario, ignoring most of the finer points of the system, would be a boring slog where half of the turns are uneventful for 5 turns until 0HP is reached.